Add ACOG to drop list
Adds ACOG scope to drop list in the gunmod_milspec group to maintain balance.

I assume the number after the item is spawn chance, so I copied the holo here.  If this is not the case, or if this for some reason breaks game balance please change and let me know.